refactor(applications): employee role and pending removal state
