+++ backend/src/modules/application/dto/organization-application.filters.dto.ts
@@ -1,12 +1,11 @@
-import { IsEnum, IsNumber, IsOptional } from 'class-validator';
-import { ApplicationStatus } from '../enums/application-status.enum';
+import { IsNumber, IsOptional } from 'class-validator';
 
 export class OrganizationApplicationFilterDto {
   @IsOptional()
-  @IsEnum(ApplicationStatus)
-  status?: ApplicationStatus;
+  @IsNumber()
+  organizationId?: number;
 
   @IsOptional()
   @IsNumber()
-  organizationId?: number;
+  userId?: number;
 }
